refactor: change intents to commands and add keybindings
