+++ src/components/App/globals.tsx
@@ -9,6 +9,7 @@ import { useAppRouter } from "./hooks/routing";
 import { useTitleSync } from "./hooks/title";
 import { useViewSync } from "./hooks/view";
 import { useWindowSettings } from "./hooks/window";
+import { useMenuBar } from "./hooks/menu";
 
 export function Globals(): ReactNode {
 	useModKeyTracker();
@@ -22,6 +23,7 @@ export function Globals(): ReactNode {
 	useAppRouter();
 	useIntercom();
 	usePolicyAlert();
+	useMenuBar();
 
 	return;
 }
